Get node Cinder device path from /dev/disk/by-id
The information provided by the OpenStack API for a Cinder volume's
device path on a VM can't be trusted in all cases. Rather than trust the
API, try to find the device path by matching the volume's UUID against
file names in /dev/disk/by-id.

The function that gets the device path is slightly modified from the
GetDevicePathBySerialID method of openstack_volumes.go here:

pkg/cloudprovider/providers/openstack/openstack_volumes.go

The function returns an error in case it can't find a matching device
path, as we want to consider this to be a fatal error.

It looks like the .travis.yml config no longer works, so remove it for
now.

It looks like:

- Travis is no longer used by upstream
- The current Travis configuration does not work due to reasons
  unrelated to this patch

Since we're patching an outdated v0.3.0 version and any patch submitted
upstream would be made against the current master branch, let's remove
.travis.yml as fixing it would not provide much benefit for this
(hopefully) one-off patch.